add makefile, dockerfile, support for yml config
