+++ .dockerignore
@@ -1,0 +1,21 @@
+**/*.pyc
+**/local_settings.py
+**/.idea
+analytics.conf
+**/resources/*
+**/media/*
+media/
+**/collected_static/
+**/env
+**/core/static/bower_components
+**/node_modules
+static
+**/env
+**/webpack-stats.json
+**/npm-debug.log
+**/assets
+**/yarn-error.log
+**/pdf/
+**/images/
+**/prove.conf
+**/.vscode
