+++ .gitignore
@@ -1,3 +1,6 @@
 .idea/*
+.DS_Store
+.vscode
 server
 main
+vendor
