+++ docker.yml
@@ -1,0 +1,6 @@
+---
+
+ip: 0.0.0.0
+port: 8888
+domain: 'softree.group'
+secure: True
