+++ go.mod
@@ -3,7 +3,7 @@ module server
 go 1.15
 
 require (
-	github.com/Finnhub-Stock-API/finnhub-go v1.2.1 // indirect
 	github.com/gorilla/handlers v1.5.1
 	github.com/gorilla/mux v1.8.0
+	gopkg.in/yaml.v2 v2.3.0
 )
