+++ local.yml
@@ -1,0 +1,6 @@
+---
+
+ip: 127.0.0.1
+port: 8888
+domain: 'self.ru'
+secure: True
