+++ src/domain/entity/jsonRealisation/errorJSON.go
