+++ src/domain/entity/jsonRealisation/jsonInterface.go
