+++ src/domain/entity/jsonRealisation/loginJSON.go
