+++ src/domain/entity/jsonRealisation/userJSON.go
