+++ src/handlers/authorization/login/login.go
@@ -1,0 +1,47 @@
+package login
+
+import (
+	"encoding/json"
+	"net/http"
+	"server/src/domain/entity/jsonRealisation"
+	"server/src/handlers/authorization/auth"
+	"server/src/handlers/authorization/utils"
+	"server/src/infrastructure/security"
+)
+
+func Login(w http.ResponseWriter, r *http.Request) {
+	if r.Method == http.MethodOptions {
+		return
+	}
+	defer r.Body.Close()
+
+	var loginJSON jsonRealisation.LoginJSON
+	if err := loginJSON.FillFields(r.Body); err != nil {
+		w.WriteHeader(http.StatusBadRequest)
+		return
+	}
+
+	if _, exists := utils.UsersServerSession[loginJSON.Email]; !exists {
+		w.WriteHeader(http.StatusBadRequest)
+		return
+	}
+	if utils.UsersServerSession[loginJSON.Email] != security.MakeShieldedHash(loginJSON.Password) {
+		w.WriteHeader(http.StatusBadRequest)
+		return
+	}
+
+	cookie := security.MakeCookie()
+	utils.Sessions[loginJSON.Email] = cookie.Value
+	http.SetCookie(w, &cookie)
+
+	u := auth.FindUserInSession(cookie.Value)
+	result, err := json.Marshal(&u)
+	if err != nil {
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+
+	w.Header().Set("content-type", "application/json")
+	w.WriteHeader(http.StatusOK)
+	w.Write(result)
+}
